updated p21 import scripts
